feat: add a configuration option `sourceLocale`
